pyproject.toml
[tool.poetry]
name = "quiz-correcter"
version = "0.1.0"
description = ""
authors = ["Timon Viola <44016238+timonviola@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/violatimon/quiz_correcter"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

[tool.poetry.scripts]
qco = "quiz_correcter.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"
rich-click = "^1.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "0.4.4"

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.ruff.lint.per-file-ignores]
"airflow/__init__.py" = ["F401"]
